history: O(1) time complexity when pressing up or down key
Achieve O(1) time complexity when pressing the up down down key.
The time complexity of the previous implementation is O(n) because
it maintains an index to find the corresponding entry. Instead, this
commit maintains a in-memory cache to keep the currently selected
history entry.

This approch is inspired by "picking the next task" of Linux
CFS (Complete Fair Scheduler). CFS has the cached "rb_leftmost" to
pick the next task with O(1) time complexity.